% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergent_regions.R
\name{estimate_convergent_regimes}
\alias{estimate_convergent_regimes}
\title{Detects convergent regimes under an OU model}
\usage{
estimate_convergent_regimes(model, criterion = c("AICc", "pBIC", "BIC"),
  method = c("backward", "rr"))
}
\arguments{
\item{model}{object of class l1ou returned by \code{\link{estimate_shift_configuration}}.}

\item{criterion}{information criterion for model selection (see Details in \code{\link{configuration_ic}}).}

\item{method}{method for finding convergent regimes. ``rr'' is based on genlasso, a regularized linear regression estimation. ``backward'' is a heuristic method similar to \code{surface_backward} function which is from the \code{surface} package. \code{surface_backward} function is to carry out the backward phase of SURFACE’s stepwise AIC routine}
}
\description{
This  function  takes  in  one  or  multiple  traits,  and  automatically  finds convergent regimes given estimated evolutionary shift positions returned by \code{\link{estimate_shift_configuration}}. The model assumes an Ornstein-
Uhlenbeck process whose parameters are estimated (adaptation ‘strength’ and drift variance2).
}
\examples{

library(l1ou)
data("lizard.traits", "lizard.tree")
Y <- lizard.traits[, 1:1]

fit_ind <- estimate_shift_configuration(lizard.tree, Y, criterion="AICc")
fit_conv <- estimate_convergent_regimes(fit_ind, criterion = "AICc",method="backward")
fit_ind
fit_conv
plot(fit_ind)
plot(fit_conv)

}
\seealso{
\code{\link{estimate_shift_configuration}}
}

