% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convergent_regions.R
\name{estimate_convergent_regions}
\alias{estimate_convergent_regions}
\title{Detects convergent regions under an OU model}
\usage{
estimate_convergent_regions(model, criterion = c("AICc", "pBIC"))
}
\arguments{
\item{model}{object of class l1ou returned by \code{\link{estimate_shift_configuration}}.}

\item{criterion}{information criterion for model selection (see Details in \code{\link{configuration_ic}}).}
}
\description{
Given estimated evolutionary shifts this function finds convergent regions
}
\examples{
library(l1ou)
data("lizard.traits", "lizard.tree")
Y <- lizard.traits[, 1:1]

eModel <- estimate_shift_configuration(lizard.tree, Y, criterion = "AICc")
eModel <- estimate_convergent_regions(eModel, criterion="AICc")
plot(eModel)
}

