% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_shift_configuration}
\alias{get_shift_configuration}
\title{Returns the best shift configuration with a given number of shifts among the shift configurations that have been evaluated.}
\usage{
get_shift_configuration(model, nShifts)
}
\arguments{
\item{model}{object of class l1ou returned by \code{\link{estimate_shift_configuration}}.}

\item{nShifts}{number of shifts.}
}
\value{
indices of the edges with shifts
}
\description{
Returns the best shift configuration with a given number of shifts among the shift configurations that have been evaluated.
}

