% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{l1ou_plot_phylo}
\alias{l1ou_plot_phylo}
\title{plots the tree and trait(s)}
\usage{
l1ou_plot_phylo(tree, model, ...)
}
\arguments{
\item{tree}{a phylogenetic tree of class phylo.}

\item{model}{the returned object from \code{\link{estimate_shift_configuration}}.}

\item{...}{further arguments to be passed on to plot.phylo}
}
\value{
none.
}
\description{
plots the tree and trait(s)
}
\details{
the results of sequential and parallel runs are not necessary equal.
}
\examples{
data("lizard.traits", "lizard.tree")
Y <- lizard.traits[,1]
eModel <- estimate_shift_configuration(lizard.tree, Y)
ew <- rep(1,198) # the tree has 198 edges
ew[eModel$shift.configuration] <- 3
l1ou_plot_phylo(lizard.tree, eModel, cex=0.5, label.offset=0.02, edge.width=ew)
}

