% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lizard.traits.R
\docType{data}
\name{lizard.traits}
\alias{lizard.traits}
\title{Morphospace of 100 Anolis lizards on Caribbean islands}
\format{A matrix with 100 rows and 4 columns}
\usage{
data(lizard.traits)
}
\description{
Morphospace of 100 Anolis lizards on Caribbean islands
}
\references{
Mahler, D. Luke, et al. "Exceptional convergence on the macroevolutionary landscape in island lizard radiations." Science 341.6143 (2013): 292-295.
}
\keyword{datasets}

