% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lizard.tree.R
\docType{data}
\name{lizard.tree}
\alias{lizard.tree}
\title{Phylogenetic tree of 100 Anolis lizards on Caribbean islands}
\format{a phylogenetic tree of class phylo with 100 species.}
\usage{
data(lizard.tree)
}
\description{
Phylogenetic tree of 100 Anolis lizards on Caribbean islands
}
\references{
Mahler, D. Luke, et al. "Exceptional convergence on the macroevolutionary landscape in island lizard radiations." Science 341.6143 (2013): 292-295.'
}
\keyword{datasets}

