% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shift_configuration.R
\name{model_ic}
\alias{model_ic}
\title{compute the information criterion score for the given configuration}
\usage{
model_ic(tr, Y, shift.configuration, criterion = c("pBIC", "pBICess", "mBIC",
  "BIC", "AIC", "AICc"), root.model = c("OUrandomRoot", "OUfixedRoot"))
}
\arguments{
\item{tr}{the input phylogeny.}

\item{Y}{the trait vector/matrix where it is labeled by the species names appear as row names.}

\item{shift.configuration}{the position of the shifts.}

\item{criterion}{the information criterion.}

\item{root.model}{the asncestoral state model.}
}
\value{
returns the score of the input model
}
\description{
compute the information criterion score for the given configuration
}
\examples{
library("l1ou");
data("lizardTraits", "lizardTree");
Y      <- lizard.traits[,1];
eModel <- est_shift_configuration(lizard.tree, Y);
ic.score  <- model_ic(lizard.tree, eModel$Y, eModel$shift.configuration, criterion="pBIC");
print(ic.score);
}

