% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{normalize_tree}
\alias{normalize_tree}
\title{Normalizes branch lengths to a unit tree height}
\usage{
normalize_tree(tree)
}
\arguments{
\item{tree}{ultrametric tree of class phylo with branch lengths, and edges in postorder.}
}
\value{
normalized phylogenetic tree, of class phylo.
}
\description{
Normalizes all branch lengths by the same factor, so that the distance from the root to all tips is equal to one.
}

