% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{print.l1ou}
\alias{print.l1ou}
\title{Prints out a summary of the model}
\usage{
\method{print}{l1ou}(model, ...)
}
\arguments{
\item{model}{object of class l1ou returned by \code{\link{estimate_shift_configuration}}.}

\item{...}{further arguments.}
}
\value{
none.
}
\description{
prints out a summary of the model
}
\examples{
data(lizard.traits, lizard.tree)
Y <- lizard.traits[,1]
eModel <- estimate_shift_configuration(lizard.tree, Y)
print(eModel)
}

