% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{profile.l1ou}
\alias{profile.l1ou}
\title{Prints out a summary of the shift configurations investigated by \code{\link{estimate_shift_configuration}}}
\usage{
\method{profile}{l1ou}(model, ...)
}
\arguments{
\item{model}{object of class l1ou returned by \code{\link{estimate_shift_configuration}}.}

\item{...}{further arguments.}
}
\value{
\item{shift.configurations}{list of shift configurations sorted by number of shifts.}
\item{scores}{list of scores corresponding to shift.configurations.}
\item{nShifts}{number of shifts corresponding to the shift configurations.}
}
\description{
prints the list of the shift configurations sorted by number of shifts and corresponding ic scores.
}
\examples{
data(lizard.traits, lizard.tree)
Y <- lizard.traits[,1]
eModel <- estimate_shift_configuration(lizard.tree, Y)
model.profile  <- profile(eModel)
plot(model.profile$nShifts, model.profile$scores)
}

