% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqrt_OU_covariance.R
\name{sqrt_OU_covariance}
\alias{sqrt_OU_covariance}
\title{computes the nagative square root and square root of the phylogeny covaiance matrix.}
\usage{
sqrt_OU_covariance(tr, alpha = 0, root.model = c("OUrandomRoot",
  "OUfixedRoot"))
}
\arguments{
\item{tr}{the input ultrametric phylogeny.}

\item{alpha}{the adaptation rate for OU model.}

\item{root.model}{the model of phylogeny ancestoral state.}
}
\value{
returns negative square root and square root of the phylogeny covariance matrix, i.e. \eqn{ \Sigma^{-1/2} } and \eqn{ \Sigma^{1/2}}.

library("l1ou");
data("lizardTree");
res <- sqrt_OU_covariance(lizard.tree);
}
\description{
computes the nagative square root and square root of the phylogeny covaiance matrix.
}

