% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{adjust_data}
\alias{adjust_data}
\title{Adjusts the tree and traits to meet the requirements of \code{estimate_shift_configuration}}
\usage{
adjust_data(tree, Y, normalize = TRUE, quietly = FALSE)
}
\arguments{
\item{tree}{ultrametric tree of class phylo with branch lengths.}

\item{Y}{trait vector/matrix without missing entries.}

\item{normalize}{logical. If TRUE, normalizes branch lengths to a unit tree height.}

\item{quietly}{logical. If FALSE, changes in tree/trait are printed.}
}
\value{
\item{tree}{tree of class phylo, with the same topology as the input \code{tree} but adjusted edge order.}
\item{Y}{trait vector/matrix with adjusted row names and row order.}
}
\description{
Creates a new tree and new data matrix, where the tree edges are in postorder,
the trait matrix has tip labels as row names, and the tips come in the same order
in the trait matrix as in the tree's tip.label vector.
}
\examples{
data(lizard.tree, lizard.traits)
lizard <- adjust_data(lizard.tree, lizard.traits[,1])
}

