% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{adjust_data}
\alias{adjust_data}
\title{Adjusts the tree and trait values so that they pass the required conditions of \code{estimate_shift_configuration} function conditions}
\usage{
adjust_data(tree, Y, normalize = TRUE, quietly = FALSE)
}
\arguments{
\item{tree}{ultrametric tree of class phylo with branch lengths.}

\item{Y}{trait vector/matrix without missing entries.}

\item{normalize}{logical. If TRUE, normalizes branch lengths to a unit tree height.}

\item{quietly}{logical. If FALSE, changes in tree/trait are printed.}
}
\value{
\item{tree}{the adjusted tree.}
\item{Y}{the adjusted trait vector/matrix.}
}
\description{
Changes the order of the tree to postorder. Reorders/adds the trait names so that it matches tree tip names in same order.
}
\examples{
data(lizard.tree, lizard.traits)
lizard <- adjust_data(lizard.tree, lizard.traits[,1])
}

