% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylogeny_bootstrap.R
\name{bootstrap_support}
\alias{bootstrap_support}
\title{computes the bootstrap support for the detected shift positions}
\usage{
bootstrap_support(tr, model, nItrs = 100, multicore = FALSE, nCores = 2)
}
\arguments{
\item{tr}{The input phylogeny.}

\item{model}{It contains estimated shift positions and also the input configuration. You may want to change model$opt to run with different options.}

\item{nItrs}{Maximum number of shifts; The default value is half the number of tips.}

\item{multicore}{If TRUE, it runs nCores processes in parallel. See details.}

\item{nCores}{The desire number of parallel process.}
}
\value{
detection rate vector.
}
\description{
computes the bootstrap support for the detected shift positions
}
\details{
The results of sequential and parellel runs are not necessarly equal.
}
\examples{
library("l1ou");
data("lizardTraits", "lizardTree");
Y         <- lizard.traits[,1];
eModel    <- estimate_shift_configuration(lizard.tree, Y);
result    <- bootstrap_support(lizard.tree, eModel, nItrs=2);
print(result);
}

