% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shift_configuration.R
\name{configuration_ic}
\alias{configuration_ic}
\title{compute the information criterion score for the given configuration}
\usage{
configuration_ic(tr, Y, shift.configuration, criterion = c("pBIC", "pBICess",
  "mBIC", "BIC", "AIC", "AICc"), root.model = c("OUrandomRoot",
  "OUfixedRoot"))
}
\arguments{
\item{tr}{an ultrametric phylogenetic tree of type phylo with branch lengths.}

\item{Y}{the trait vector/matrix where it is labeled by the species names appear as row names.}

\item{shift.configuration}{the shift positions, i.e. indicies of edges where the estimated shifts occurs.}

\item{criterion}{the information criterion.}

\item{root.model}{the asncestoral state model.}
}
\value{
the information criterion value of the shift configuration.
}
\description{
compute the information criterion score for the given configuration
}
\examples{
library("l1ou");
data("lizard.traits", "lizard.tree");
Y <- lizard.traits[,1];
eModel <- estimate_shift_configuration(lizard.tree, Y);
ic.score <- configuration_ic(lizard.tree, eModel$Y, eModel$shift.configuration, criterion="pBIC");
print(ic.score);
}
\seealso{
\code{\link{estimate_shift_configuration}}
}

