% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shift_configuration.R
\name{configuration_ic}
\alias{configuration_ic}
\title{computes the information criterion score for a given configuration}
\usage{
configuration_ic(tree, Y, shift.configuration, criterion = c("pBIC",
  "pBICess", "mBIC", "BIC", "AIC", "AICc"), root.model = c("OUrandomRoot",
  "OUfixedRoot"))
}
\arguments{
\item{tree}{ultrametric tree of class phylo, with branch lengths, and edges in postorder.}

\item{Y}{trait vector/matrix without missing entries. The row names of the data must be in the same order as the tip labels.}

\item{shift.configuration}{shift positions, i.e. vector of indices of the edges where the shifts occur.}

\item{criterion}{an information criterion (see Details).}

\item{root.model}{an ancestral state model at the root.}
}
\value{
Information criterion value of the given shift configuration.
}
\description{
computes the information criterion score for a given configuration
}
\details{
AIC gives the usual Akaike information criterion, penalizing each parameter by 2, and counting each shift as 2 parameter (for the shift magnitude for the shift position, as if this position were a continuous parameter).
AICc gives the usual small-sample size modification to AIC.
BIC also gives the usual Bayesian information criterion, here penalizing each shift as 2 parameters.
mBIC is the modified BIC proposed by Ho and Ané (2014).
pBIC is the phylogenetic BIC for shifts proposed by Khabbazian et al.
pBICess is a version of pBIC where the determinant term is replaced by a sum of the log of effective sample sizes (ESS), similar to the ESS proposed by Ané (2008).
}
\examples{
library("l1ou")
data("lizard.traits", "lizard.tree")
Y <- lizard.traits[,1]
eModel <- estimate_shift_configuration(lizard.tree, Y)
ic.score <- configuration_ic(lizard.tree, eModel$Y, eModel$shift.configuration, criterion="pBIC")
print(ic.score)
}
\references{
Cécile Ané, 2008. "Analysis of comparative data with hierarchical autocorrelation". Annals of Applied Statistics 2(3):1078-1102.
Ho, L. S. T. and Ané, C. 2014.  "Intrinsic inference difficulties for trait evolution with Ornstein-Uhlenbeck models". Methods in Ecology and Evolution. 5(11):1133-1146.
Mohammad Khabbazian, Ricardo Kriebel, Karl Rohe, and Cécile Ané. "Fast and accurate detection of evolutionary shifts in Ornstein-Uhlenbeck models". In review.
}
\seealso{
\code{\link{estimate_shift_configuration}}
}

