% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shift_configuration.R
\name{estimate_shift_configuration}
\alias{estimate_shift_configuration}
\title{detectes evolutionary shifts}
\usage{
estimate_shift_configuration(tr, Y,
  max.nShifts = floor(length(tr$tip.label)/2), criterion = c("pBIC",
  "pBICess", "mBIC", "BIC", "AIC", "AICc"), root.model = c("OUrandomRoot",
  "OUfixedRoot"), quietly = TRUE, alpha.upper = alpha_upper_bound(tr),
  alpha.lower = 0, standardize = TRUE,
  num.top.configurations = max.nShifts/2, edge.length.threshold = 10 *
  .Machine$double.eps, grp.delta = 1/16, grp.seq.ub = 5,
  l1ou.options = NA)
}
\arguments{
\item{tr}{the input phylogeny. It should be an ultermetric tree in postorder.}

\item{Y}{the trait vector/matrix where it is labeled by the species names appear as row names. The rownames must match the tips label with the same order.}

\item{max.nShifts}{maximum number of shifts; The default value is half the number of tips.}

\item{criterion}{the type if information criterion for model selection.}

\item{root.model}{the ancestral state model.}

\item{quietly}{logical. If FALSE, it writes to the output.}

\item{alpha.upper}{the upper bound value of phylogenetic adaptation rate for computing maximum likelihood estimation. By default it is log(2) over the minimum length of branches connected to tips (they are supposed to be non-zero).}

\item{alpha.lower}{lower bound value of phylogenetic adaptation rate.}

\item{standardize}{logical. If TRUE, the columns of the trait matrix will be standardized.}

\item{num.top.configurations}{an internal argument. It is the number of good shift configuration that is chosen for further improvement.}

\item{edge.length.threshold}{the minim edge length that is considered non-zero.}

\item{grp.delta}{an internal parameter. The input lambda sequence for grplasso will be lamda.max*(0.5^ (0, grp.seq.ub, grp.delta) ).}

\item{grp.seq.ub}{an internal parameter. The input lambda sequence for grplasso will be lamda.max*(0.5^ (0, grp.seq.ub, grp.delta) ).}

\item{l1ou.options}{if the option object is provided, all the default values will be ignored. It is good for the bootstrap procedure to be run with previously used options.}
}
\value{
returns estimated model.
}
\description{
detectes evolutionary shifts
}
\examples{
data("lizard.traits", "lizard.tree");
Y = lizard.traits[,1];
eModel <- estimate_shift_configuration(lizard.tree, Y);
ew <- rep(1, 198) # the tree has 198 edges
ew[eModel$shift.configuration] <- 3
l1ou_plot_phylo(lizard.tree, eModel, "PC1", cex=0.5, label.offset=0.02, edge.width=ew);
}

