% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shift_configuration.R
\name{estimate_shift_configuration}
\alias{estimate_shift_configuration}
\title{Detects evolutionary shifts under an OU model}
\usage{
estimate_shift_configuration(tree, Y,
  max.nShifts = floor(length(tree$tip.label)/2), criterion = c("pBIC",
  "pBICess", "mBIC", "BIC", "AIC", "AICc"), root.model = c("OUfixedRoot",
  "OUrandomRoot"), candid.edges = NA, quietly = TRUE,
  alpha.upper = alpha_upper_bound(tree), alpha.lower = NA,
  standardize = TRUE, num.top.configurations = max.nShifts/2,
  edge.length.threshold = .Machine$double.eps, grp.delta = 1/16,
  grp.seq.ub = 5, l1ou.options = NA)
}
\arguments{
\item{tree}{ultrametric tree of class phylo with branch lengths, and edges in postorder.}

\item{Y}{trait vector/matrix without missing entries. The row names of the data must be in the same order as the tip labels.}

\item{max.nShifts}{upper bound for the number of shifts. The default value is half the number of tips.}

\item{criterion}{information criterion for model selection (see Details in \code{\link{configuration_ic}}).}

\item{root.model}{ancestral state model at the root.}

\item{candid.edges}{a vector of indices of candidate edges where the shifts may occur. If provided, shifts will only be allowed on these edges; otherwise all edges will be considered.}

\item{quietly}{logical. If FALSE, a basic summary of the progress and results is printed.}

\item{alpha.upper}{upper bound for the phylogenetic adaptation rate. The default value is log(2) over the minimum branch length connected to tips.}

\item{alpha.lower}{lower bound for the phylogenetic adaptation rate.}

\item{standardize}{logical. If TRUE, the columns of the trait matrix are first standardized.}

\item{num.top.configurations}{number of shift configurations on the lasso solution path chosen for further improvement.}

\item{edge.length.threshold}{minimum edge length that is considered non-zero. Branches with shorter length are considered as soft polytomies, disallowing shifts on such branches.}

\item{grp.delta}{internal (used when the data contain multiple traits). The input lambda sequence for the group lasso, in `grplasso', will be lambda.max*(0.5^seq(0, grp.seq.ub, grp.delta) ).}

\item{grp.seq.ub}{(used for multiple traits). The input lambda sequence for grplasso will be lambda.max*(0.5^seq(0, grp.seq.ub, grp.delta) ).}

\item{l1ou.options}{if provided, all the default values will be ignored.}
}
\value{
\item{Y}{input trait vector/matrix.}
\item{shift.configuration}{estimated shift positions, i.e. vector of indices of edges where the estimated shifts occur.}
\item{shift.values}{estimates of the shift values.}
\item{nShifts}{estimated number of shifts.}
\item{optimums}{optimum values of the trait along the edges. If the data are multivariate, this is a matrix where each row corresponds to an edge.}
\item{alpha}{maximum likelihood estimate(s) of the adaptation rate \eqn{\alpha}{alpha}, one per trait.}
\item{sigma2}{maximum likelihood estimate(s) of the variance rate \eqn{\sigma^2}{sigma^2}, one per trait.}
\item{mu}{fitted values, i.e. estimated trait means.}
\item{residuals}{residuals. These residuals are phylogenetically correlated.}
\item{score}{information criterion value of the estimated shift configuration.}
\item{l1ou.options}{list of options that were used.}
}
\description{
This function takes in one or multiple traits, and automatically detects the phylogenetic placement and
the magnitude of shifts in the evolution of these traits. The model assumes an Ornstein-Uhlenbeck process
whose parameters are estimated (adaptation `strength' \eqn{\alpha}{alpha} and drift variance \eqn{\sigma^2}{sigma^2}).
Instantaneous shifts in the optimal trait value affect the traits over time.
}
\details{
For information criteria: see \code{\link{configuration_ic}}.
}
\examples{
data(lizard.tree, lizard.traits)
lizard <- adjust_data(lizard.tree, lizard.traits[,1])
eModel <- estimate_shift_configuration(lizard$tree, lizard$Y)
nEdges <- length(lizard.tree$edge[,1]) # total number of edges
ew <- rep(1,nEdges)                    # to set default edge width of 1
ew[eModel$shift.configuration] <- 3    # to widen edges with a shift
plot_l1ou(lizard$tree, eModel, cex=0.5, label.offset=0.02, edge.width=ew)

# example to constrain the set of candidate branches with a shift
eModel <- estimate_shift_configuration(lizard$tree, lizard$Y, criterion="AICc")
ce <- eModel$shift.configuration # set of candidate edges
eModel <- estimate_shift_configuration(lizard$tree, lizard$Y, candid.edges = ce)
plot_l1ou(lizard$tree, eModel, edge.ann.cex=0.7, cex=0.5, label.offset=0.02)
}
\references{
Mohammad Khabbazian, Ricardo Kriebel, Karl Rohe, and Cécile Ané. "Fast and accurate detection of evolutionary shifts in Ornstein-Uhlenbeck models". In review.
}

