% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{l1ou_plot_phylo}
\alias{l1ou_plot_phylo}
\title{plots the tree and trait(s)}
\usage{
l1ou_plot_phylo(tr, model, title.str = paste(1:ncol(model$Y)),
  enable.cross = FALSE, ...)
}
\arguments{
\item{tr}{the input phylogeny.}

\item{model}{it contains estimated shift positions and also the input configuration. You may want to change model$opt to run with different options.}

\item{title.str}{the title for each trait.}

\item{enable.cross}{logical. If TRUE, annotates each shift by X.}

\item{...}{extra parameters for plot.phylo}
}
\description{
plots the tree and trait(s)
}
\details{
the results of sequential and parallel runs are not necessary equal.
}
\examples{
data("lizard.traits", "lizard.tree");
Y <- lizard.traits[,1];
eModel <- estimate_shift_configuration(lizard.tree, Y);
ew <- rep(1,198) # the tree has 198 edges
ew[eModel$shift.configuration] <- 3
l1ou_plot_phylo(lizard.tree, eModel, "PC1",cex=0.5, label.offset=0.02, edge.width=ew);
}

