% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{normalize_tree}
\alias{normalize_tree}
\title{normalizes the branch lengths so that the distance from the root to all tips are equal to one.}
\usage{
normalize_tree(tr)
}
\arguments{
\item{tr}{an ultrametric phylogenetic tree of type phylo with branch lengths.}
}
\value{
the normalized phylogeny tree.
}
\description{
normalizes the branch lengths so that the distance from the root to all tips are equal to one.
}

