% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.l1ou}
\alias{plot.l1ou}
\title{Visualizes a shift configuration: tree and trait(s)}
\usage{
\method{plot}{l1ou}(model, palette = NA, edge.shift.ann = TRUE,
  edge.shift.adj = c(0.5, -0.025), edge.label = c(), star = TRUE,
  edge.label.ann = FALSE, edge.label.adj = c(0.5, 1), edge.label.pos = NA,
  edge.ann.cex = 1, plot.bar = TRUE, bar.axis = TRUE, ...)
}
\arguments{
\item{model}{object of class l1ou returned by \code{\link{estimate_shift_configuration}}.}

\item{palette}{vector of colors, of size the number of shifts plus one. The last element is the color for the background regime (regime at the root).}

\item{edge.shift.ann}{logical. If TRUE, annotates edges by shift values.}

\item{edge.shift.adj}{adjustment argument to give to edgelabel() for labeling edges by shift values.}

\item{edge.label}{vector of size number of edges.}

\item{star}{logical. If TRUE, the shift positions will be annotated by "*". It is useful for gray scale plots.}

\item{edge.label.ann}{logical. If TRUE, annotates edges by labels in tree$edge.label, if non-empty, or edge.label.}

\item{edge.label.adj}{adjustment argument to give to edgelabel() for labeling edges.}

\item{edge.label.pos}{relative position of the edge.label on the edge. 0 for the beginning of the edge and 1 for the end of the edge.}

\item{edge.ann.cex}{amount by which the annotation text should be magnified relative to the default.}

\item{plot.bar}{logical. If TRUE, the bars corresponding to the trait values will be plotted.}

\item{bar.axis}{logical. If TRUE, the axis of of trait(s) range will be plotted.}

\item{...}{further arguments to be passed on to plot.phylo.}
}
\value{
none.
}
\description{
plots the tree annotated to show the edges with a shift, and the associated trait data side by side.
}
\examples{
data(lizard.traits, lizard.tree)
Y <- lizard.traits[,1]
eModel <- estimate_shift_configuration(lizard.tree, Y)
nEdges <- length(lizard.tree$edge[,1])
ew <- rep(1,nEdges)
ew[eModel$shift.configuration] <- 3
plot(eModel, cex=0.5, label.offset=0.02, edge.width=ew)
}

