% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_l1ou}
\alias{plot_l1ou}
\title{plots the tree and trait(s)}
\usage{
plot_l1ou(tree, model, pallet = NA, bar.axis = TRUE, ...)
}
\arguments{
\item{tree}{a phylogenetic tree of class phylo.}

\item{model}{the returned object from \code{\link{estimate_shift_configuration}}.}

\item{pallet}{a color vector of size number of the shifts plus one. The last element is the background color.}

\item{bar.axis}{logical. If TRUE, it plots axis for trait(s) range.}

\item{...}{further arguments to be passed on to plot.phylo}
}
\value{
none.
}
\description{
plots the tree and trait(s)
}
\examples{
data("lizard.traits", "lizard.tree")
Y <- lizard.traits[,1]
eModel <- estimate_shift_configuration(lizard.tree, Y)
ew <- rep(1,198) # the tree has 198 edges
ew[eModel$shift.configuration] <- 3
plot_l1ou(lizard.tree, eModel, cex=0.5, label.offset=0.02, edge.width=ew)
}

