% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqrt_OU_covariance.R
\name{sqrt_OU_covariance}
\alias{sqrt_OU_covariance}
\title{computes the negative square root and square root of the phylogeny covariance matrix in linear time complexity.}
\usage{
sqrt_OU_covariance(tree, alpha = 0, root.model = c("OUfixedRoot",
  "OUrandomRoot"))
}
\arguments{
\item{tree}{an ultrametric phylogenetic tree of class phylo with branch lengths.}

\item{alpha}{the adaptation rate for the OU model.}

\item{root.model}{an ancestral state model at the root.}
}
\value{
\item{sqrtSigma}{square root of the phylogeny covariance matrix.}
\item{sqrtInvSigma}{inverse square root of the phylogeny covariance matrix.}
}
\description{
computes the negative square root and square root of the phylogeny covariance matrix in linear time complexity.
}
\examples{
library("l1ou")
data("lizard.tree")
res <- sqrt_OU_covariance(lizard.tree)
Sigma <- vcv(lizard.tree)
dimnames(Sigma) <- NULL
stopifnot( all.equal(res$sqrtSigma \%*\% t(res$sqrtSigma) , Sigma) ) # TRUE
stopifnot( all.equal(res$sqrtInvSigma \%*\% t(res$sqrtInvSigma) , solve(Sigma)) ) # TRUE
}
\references{
M. Khabbazian, R. Kriebel, K. Rohe, and Cecile Ane. Fast and accurate detection of evolutionary shifts in Ornstein-Uhlenbeck models
}

