% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqrt_OU_covariance.R
\name{sqrt_OU_covariance}
\alias{sqrt_OU_covariance}
\title{computes the nagative square root and square root of the phylogeny covaiance matrix.}
\usage{
sqrt_OU_covariance(tr, alpha = 0, root.model = c("OUfixedRoot",
  "OUrandomRoot"))
}
\arguments{
\item{tr}{an ultrametric phylogenetic tree of type phylo with branch lengths.}

\item{alpha}{the adaptation rate for OU model.}

\item{root.model}{the model of phylogeny ancestoral state.}
}
\value{
\item{sqrtSigma}{square root of phylogeny covariance matrix}.
\item{sqrtInvSigma}{inverse square root of phylogeny covariance matrix.}
}
\description{
computes the nagative square root and square root of the phylogeny covaiance matrix.
}
\examples{
library("l1ou")
data("lizard.tree")
res <- sqrt_OU_covariance(lizard.tree);
Sigma <- vcv(lizard.tree)
dimnames(Sigma) <- NULL
all.equal(res$sqrtSigma \%*\% t(res$sqrtSigma) , Sigma) # TRUE
all.equal(res$sqrtInvSigma \%*\% t(res$sqrtInvSigma) , solve(Sigma)) # TRUE
}

